+++ CHANGELOG.md
@@ -245,6 +245,16 @@ Version 1.3.1 - some new features, Refactors, bug fixes and many enhancements.
 - [x] fixed UI text case 
 - [x] rafactor style settings, fix style managers, remove style settings code from hoverpopover.
 
+
+### Version 1.3.2
+
+Version 1.3.2 - some new features, enhancements and bug fixes.
+
+⭐ New Features :
+- [x] Add [feature #104](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/104), add a button to forbid equation panel to be refreshed.
+- [x] Add [feature #105](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/105). Add a settings in auto-numbering settings tab, only number equations already with tags.
+
+
 ### 🚢 Comming Soon : 
 1. Auto-number and citation complete add for tables and figures 
 2. Check citations of figures and tables, all in one 😄